Crate ashpd

ASHPD, acronym of Aperture Science Handheld Portal Device is a Rust & zbus wrapper of the XDG portals DBus interfaces. The library aims to provide an easy way to interact with the various portals defined per the specifications.

It provides an alternative to the C library https://github.com/flatpak/libportal.

use ashpd::{desktop::screenshot, Response, WindowIdentifier};
use zbus::fdo::Result;

async fn run() -> Result<()> {
    let identifier = WindowIdentifier::default();
    if let Ok(Response::Ok(color)) = screenshot::pick_color(identifier).await {
        println!("({}, {}, {})", color.red(), color.green(), color.blue());
    }
    Ok(())
}

use ashpd::{desktop::camera, Response, WindowIdentifier};
use zbus::fdo::Result;

async fn run() -> Result<()> {
    let identifier = WindowIdentifier::default();
    if let Ok(Response::Ok(pipewire_node_id)) = camera::stream().await {
        // Render the stream with GStreamer for example, see the demo
    }
    Ok(())
}

Optional features

Feature	Description

feature_gtk3	Implement From<Color> for gdk3::RGBA

Implement From<gtk3::Window> for WindowIdentifier

feature_gtk4	Implement From<Color> for gdk4::RGBA

Provides WindowIdentifier::from_window

Re-exports

pub use enumflags2;

pub use zbus;

pub use zvariant;

Modules

desktop
Interact with the user's desktop such as taking a screenshot, setting a background or querying the user's location.

documents
Interact with the documents store or transfer files across apps.

flatpak
Spawn commands outside the sandbox or monitor if the running application has received an update & install it.

Structs

AsyncRequestProxy
Asynchronous sibling of RequestProxy. The Request interface is shared by all portal interfaces. When a portal method is called, the reply includes a handle (i.e. object path) for a Request object, which will stay alive for the duration of the user interaction related to the method call.

AsyncSessionProxy
Asynchronous sibling of SessionProxy. The Session interface is shared by all portal interfaces that involve long lived sessions. When a method that creates a session is called, if successful, the reply will include a session handle (i.e. object path) for a Session object, which will stay alive for the duration of the session.

BasicResponse
The most basic response. Used when only the status of the request is what we receive as a response.

HandleToken
A handle token is a DBus Object Path element, specified in the RequestProxy or SessionProxy object path following this format /org/freedesktop/portal/desktop/request/SENDER/TOKEN where sender is the caller's unique name and token is the HandleToken.

RequestProxy
The Request interface is shared by all portal interfaces. When a portal method is called, the reply includes a handle (i.e. object path) for a Request object, which will stay alive for the duration of the user interaction related to the method call.

SessionProxy
The Session interface is shared by all portal interfaces that involve long lived sessions. When a method that creates a session is called, if successful, the reply will include a session handle (i.e. object path) for a Session object, which will stay alive for the duration of the session.

Enums

Response
A typical response returned by the connect_response signal of a RequestProxy.

ResponseError
An error returned a portal request caused by either the user cancelling the request or something else.

WindowIdentifier
Most portals interact with the user by showing dialogs. These dialogs should generally be placed on top of the application window that triggered them. To arrange this, the compositor needs to know about the application window. Many portal requests expect a WindowIdentifier for this reason.

Functions

is_sandboxed
Check whether the application is running inside a sandbox.