Module ashpd::desktop::account

Request access to the current logged user information such as the id, name or their avatar uri.

Examples

use ashpd::desktop::account::AccountProxy;

async fn run() -> Result<(), ashpd::Error> {
    let connection = zbus::azync::Connection::new_session().await?;

    let proxy = AccountProxy::new(&connection).await?;
    let user_info = proxy
        .user_information(
            Default::default(),
            "App would like to access user information",
        )
        .await?;

    println!("Name: {}", user_info.name());
    println!("ID: {}", user_info.id());

    Ok(())
}

Structs

AccountProxy
The interface lets sandboxed applications query basic information about the user, like his name and avatar photo.

UserInfo
The response of a AccountProxy::user_information request.