Module ashpd::desktop::dynamic_launcher

Examples

use ashpd::{
    desktop::{
        dynamic_launcher::{DynamicLauncherProxy, LauncherType, PrepareInstallOptions},
        Icon,
    },
    WindowIdentifier,
};

async fn run() -> ashpd::Result<()> {
    let proxy = DynamicLauncherProxy::new().await?;
    let (name, token) = proxy
        .prepare_install(
            &WindowIdentifier::default(),
            "My App",
            Icon::with_names(&["dialog-symbolic"]),
            PrepareInstallOptions::default(),
        )
        .await?
        .response()?;

    // Name and Icon will be overwritten from what we provided above
    // Exec will be overridden to call `flatpak run our-app` if the application is sandboxed
    let desktop_entry = r#"
        [Desktop Entry]
        Comment=My Web App
        Type=Application
    "#;
    proxy
        .install(&token, "some_file.desktop", desktop_entry)
        .await?;

    proxy.uninstall("some_file.desktop").await?;
    Ok(())
}

Structs

DynamicLauncherProxy

The interface lets sandboxed applications install launchers like Web Application from your browser or Steam.

LauncherIcon

The icon of the launcher.

PrepareInstallOptions

Options to pass to DynamicLauncherProxy::prepare_install

Enums

IconType

The icon format.

LauncherType

The type of the launcher.