Module ashpd::desktop::location

Query the user's GPS location.

Examples

use ashpd::desktop::location::{CreateSessionOptions, LocationProxy, SessionStartOptions};
use ashpd::{BasicResponse as Basic, HandleToken, Response, WindowIdentifier};
use std::convert::TryFrom;
use zbus::{self, fdo::Result};

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = LocationProxy::new(&connection)?;

    let options = CreateSessionOptions::default()
        .session_handle_token(HandleToken::try_from("token").unwrap());

    let session = proxy.create_session(options)?;

    let request = proxy.start(
        &session,
        WindowIdentifier::default(),
        SessionStartOptions::default(),
    )?;

    request.connect_response(move |response: Response<Basic>| {
        proxy.connect_location_updated(move |location| {
            println!("{}", location.accuracy());
            println!("{}", location.longitude());
            println!("{}", location.latitude());
            Ok(())
        })?;
        Ok(())
    })?;

    Ok(())
}

Structs

AsyncLocationProxy
Asynchronous sibling of LocationProxy. The interface lets sandboxed applications query basic information about the location.

CreateSessionOptions
Specified options for a location access request.

Location
The response received on a location_updated signal.

LocationProxy
The interface lets sandboxed applications query basic information about the location.

SessionStartOptions
Specified options for a location session start request.

Enums

Accuracy
The accuracy of the location.