Module ashpd::desktop::memory_monitor

Monitor memory level.

Examples

use ashpd::desktop::memory_monitor::MemoryMonitorProxy;
use zbus::{self, fdo::Result};

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = MemoryMonitorProxy::new(&connection);
    proxy.connect_low_memory_warning(move |level| {
        println!("{:#?}", level);
        Ok(())
    })?;
    Ok(())
}

Structs

AsyncMemoryMonitorProxy
Asynchronous sibling of MemoryMonitorProxy. The interface provides information about low system memory to sandboxed applications. It is not a portal in the strict sense, since it does not involve user interaction.

MemoryMonitorProxy
The interface provides information about low system memory to sandboxed applications. It is not a portal in the strict sense, since it does not involve user interaction.