Struct ashpd::desktop::notification::NotificationProxy

pub struct NotificationProxy<'a> { /* fields omitted */ }

The interface lets sandboxed applications send and withdraw notifications.

It is not possible for the application to learn if the notification was actually presented to the user. Not a portal in the strict sense, since there is no user interaction.

Note that in contrast to most other portal requests, notifications are expected to outlast the running application. If a user clicks on a notification after the application has exited, it will get activated again.

Notifications can specify actions that can be activated by the user. Actions whose name starts with 'app.' are assumed to be exported and will be activated via the ActivateAction() method in the org.freedesktop.Application interface. Other actions are activated by sending the #org.freedeskop.portal.Notification::ActionInvoked signal to the application.

Implementations

impl<'a> NotificationProxy<'a>

pub fn new(connection: &'a Connection) -> Result<Self>

Create a new notification proxy.

pub fn on_action_invoked<F>(&self, callback: F) -> Result<()> where
F: FnOnce(Action<'_>),

Signal emitted when a particular action is invoked

pub fn add_notification(
&self,
id: &str,
notification: Notification
) -> Result<()>

Sends a notification.

The ID can be used to later withdraw the notification. If the application reuses the same ID without withdrawing, the notification is replaced by the new one.

Arguments

id - Application-provided ID for this notification

notification - HashMap

pub fn remove_notification(&self, id: &str) -> Result<()>

Withdraws a notification.

Arguments

id - Application-provided ID for this notification

pub fn version(&self) -> Result<u32>

version property