Module ashpd::desktop::remote_desktop

Start a remote desktop session and interact with it.

Examples

use ashpd::desktop::remote_desktop::{
    CreateRemoteOptions, CreateSession, DeviceType, KeyState, RemoteDesktopProxy,
    SelectDevicesOptions, SelectedDevices, StartRemoteOptions,
};
use ashpd::{BasicResponse as Basic, HandleToken, Response, WindowIdentifier};
use std::collections::HashMap;
use std::convert::TryFrom;
use zbus::{fdo::Result, Connection};
use zvariant::ObjectPath;

fn select_devices(
    handle: &'static ObjectPath<'_>,
    proxy: &'static RemoteDesktopProxy,
) -> Result<()> {
    let request = proxy.select_devices(handle,
        SelectDevicesOptions::default().types(DeviceType::Keyboard | DeviceType::Pointer),
    )?;

    request.connect_response(move |r: Response<Basic>| {
        if r.is_ok() {
            start_remote(handle, proxy)?;
        }
        Ok(())
    })?;

    Ok(())
}

fn start_remote(
    handle: &'static ObjectPath<'_>,
    proxy: &'static RemoteDesktopProxy,
) -> Result<()> {
    let request = proxy.start(
        handle,
        WindowIdentifier::default(),
        StartRemoteOptions::default(),
    )?;

    request.connect_response(move |r: Response<SelectedDevices>| {
        proxy.notify_keyboard_keycode(
            handle,
            HashMap::new(),
            13, // Enter key code
            KeyState::Pressed,
        )?;

        println!("{:#?}", r.unwrap().devices);
        Ok(())
    })?;

    Ok(())
}

fn main() -> Result<()> {
    let connection = Connection::new_session()?;
    let proxy = RemoteDesktopProxy::new(&connection);

    let request = proxy.create_session(
        CreateRemoteOptions::default()
            .session_handle_token(HandleToken::try_from("token").unwrap()),
    )?;

    request.connect_response(move |r: Response<CreateSession>| {
        let response = r.unwrap();
        select_devices(response.session_handle(), &proxy)?;
        Ok(())
    })?;

    Ok(())
}

Structs

AsyncRemoteDesktopProxy
Asynchronous sibling of RemoteDesktopProxy. The interface lets sandboxed applications create remote desktop sessions.

CreateRemoteOptions
Specified options on a create a remote session request.

CreateSession
A response to a create_session request.

RemoteDesktopProxy
The interface lets sandboxed applications create remote desktop sessions.

SelectDevicesOptions
Specified options on a select devices request.

SelectedDevices
A response to a select device request.

StartRemoteOptions
Specified options on a start remote desktop request.

Enums

Axis
The available axis.

DeviceType
A bit flag for the available devices.

KeyState
The keyboard key state.