Struct ashpd::desktop::remote_desktop::RemoteDesktopProxy

pub struct RemoteDesktopProxy<'a>(_);

The interface lets sandboxed applications create remote desktop sessions.

Wrapper of the DBus interface: org.freedesktop.portal.RemoteDesktop.

Implementations

impl<'a> RemoteDesktopProxy<'a>

pub async fn new(
connection: &Connection
) -> Result<RemoteDesktopProxy<'a>, Error>

Create a new instance of RemoteDesktopProxy.

pub fn inner(&self) -> &Proxy<'_>

Get a reference to the underlying Proxy.

pub async fn create_session(&self) -> Result<SessionProxy<'a>, Error>

Create a remote desktop session. A remote desktop session is used to allow remote controlling a desktop session. It can also be used together with a screen cast session.

pub async fn select_devices(
&self,
session: &SessionProxy<'_>,
types: BitFlags<DeviceType>
) -> Result<(), Error>

Select input devices to remote control.

Arguments

session - A SessionProxy.

types - The device types to request remote controlling of.

pub async fn start(
&self,
session: &SessionProxy<'_>,
identifier: WindowIdentifier
) -> Result<BitFlags<DeviceType>, Error>

Start the remote desktop session.

This will typically result in the portal presenting a dialog letting the user select what to share, including devices and optionally screen content if screen cast sources was selected.

Arguments

session - A SessionProxy.

identifier - The application window identifier.

pub async fn notify_keyboard_keycode(
&self,
session: &SessionProxy<'_>,
keycode: i32,
state: KeyState
) -> Result<(), Error>

Notify keyboard code. May only be called if KEYBOARD access was provided after starting the session.

Arguments

session - A SessionProxy.

keycode - Keyboard code that was pressed or released.

state - The new state of the keyboard code.

pub async fn notify_keyboard_keysym(
&self,
session: &SessionProxy<'_>,
keysym: i32,
state: KeyState
) -> Result<(), Error>

Notify keyboard symbol. May only be called if KEYBOARD access was provided after starting the session.

Arguments

session - A SessionProxy.

keysym - Keyboard symbol that was pressed or released.

state - The new state of the keyboard code.

pub async fn notify_touch_up(
&self,
session: &SessionProxy<'_>,
slot: u32
) -> Result<(), Error>

Notify about a new touch up event.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy.

slot - Touch slot where touch point appeared.

pub async fn notify_touch_down(
&self,
session: &SessionProxy<'_>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new touch down event. The (x, y) position represents the new touch point position in the streams logical coordinate space.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy.

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch down x coordinate.

y - Touch down y coordinate.

pub async fn notify_touch_motion(
&self,
session: &SessionProxy<'_>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new touch motion event. The (x, y) position represents where the touch point position in the streams logical coordinate space moved.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy.

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch motion x coordinate.

y - Touch motion y coordinate.

pub async fn notify_pointer_motion_absolute(
&self,
session: &SessionProxy<'_>,
stream: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new absolute pointer motion event. The (x, y) position represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy.

stream - The PipeWire stream node the coordinate is relative to.

x - Pointer motion x coordinate.

y - Pointer motion y coordinate.

pub async fn notify_pointer_motion(
&self,
session: &SessionProxy<'_>,
dx: f64,
dy: f64
) -> Result<(), Error>

Notify about a new relative pointer motion event. The (dx, dy) vector represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy.

dx - Relative movement on the x axis.

dy - Relative movement on the y axis.

pub async fn notify_pointer_button(
&self,
session: &SessionProxy<'_>,
button: i32,
state: KeyState
) -> Result<(), Error>

Notify pointer button. The pointer button is encoded according to Linux Evdev button codes.

May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy.

button - The pointer button was pressed or released.

state - The new state of the keyboard code.

pub async fn notify_pointer_axis_discrete(
&self,
session: &SessionProxy<'_>,
axis: Axis,
steps: i32
) -> Result<(), Error>

Notify pointer axis discrete. May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy.

axis - The axis that was scrolled.

pub async fn notify_pointer_axis(
&self,
session: &SessionProxy<'_>,
dx: f64,
dy: f64
) -> Result<(), Error>

Notify pointer axis. The axis movement from a “smooth scroll” device, such as a touchpad. When applicable, the size of the motion delta should be equivalent to the motion vector of a pointer motion done using the same advice.

May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy.

dx - Relative axis movement on the x axis.

dy - Relative axis movement on the y axis.

pub async fn available_device_types(
&self
) -> Result<BitFlags<DeviceType>, Error>

Available source types.