Struct ashpd::desktop::screencast::Screencast

pub struct Screencast<'a>(_);

The interface lets sandboxed applications create screen cast sessions.

Wrapper of the DBus interface: org.freedesktop.portal.ScreenCast.

Implementations

impl<'a> Screencast<'a>

pub async fn new() -> Result<Screencast<'a>, Error>

Create a new instance of Screencast.

pub async fn create_session(&self) -> Result<Session<'a>, Error>

Create a screen cast session.

Specifications

See also CreateSession.

pub async fn open_pipe_wire_remote(
&self,
session: &Session<'_>
) -> Result<RawFd, Error>

Open a file descriptor to the PipeWire remote where the screen cast streams are available.

Arguments

session - A Session, created with create_session().

Returns

File descriptor of an open PipeWire remote.

Specifications

See also OpenPipeWireRemote.

pub async fn select_sources(
&self,
session: &Session<'_>,
cursor_mode: CursorMode,
types: BitFlags<SourceType>,
multiple: bool,
restore_token: Option<&str>,
persist_mode: PersistMode
) -> Result<Request<()>, Error>

Configure what the screen cast session should record. This method must be called before starting the session.

Passing invalid input to this method will cause the session to be closed. An application may only attempt to select sources once per session.

Arguments

session - A Session, created with create_session().

cursor_mode - Sets how the cursor will be drawn on the screen cast stream.

types - Sets the types of content to record.

multiple- Sets whether to allow selecting multiple sources.

Specifications

See also SelectSources.

pub async fn start(
&self,
session: &Session<'_>,
identifier: &WindowIdentifier
) -> Result<Request<Streams>, Error>

Start the screen cast session.

This will typically result the portal presenting a dialog letting the user do the selection set up by select_sources.

An application can only attempt start a session once.

Arguments

session - A Session, created with create_session().

identifier - Identifier for the application window.

Return

A list of Stream and an optional restore token.

Specifications

See also Start.

pub async fn available_cursor_modes(
&self
) -> Result<BitFlags<CursorMode>, Error>

Available cursor mode.

Specifications

See also AvailableCursorModes.

pub async fn available_source_types(
&self
) -> Result<BitFlags<SourceType>, Error>

Available source types.

Specifications

See also AvailableSourceTypes.