Module ashpd::desktop::secret

Retrieve a per-application secret used to encrypt confidential data inside the sandbox.

Examples

use ashpd::desktop::secret::{RetrieveOptions, SecretProxy};
use ashpd::{BasicResponse as Basic, Response};
use std::fs::File;
use std::os::unix::io::AsRawFd;
use zbus::{fdo::Result, Connection};
use zvariant::Fd;

fn main() -> Result<()> {
    let connection = Connection::new_session()?;
    let proxy = SecretProxy::new(&connection);

    let file = File::open("test.txt").unwrap();

    let request = proxy.retrieve_secret(Fd::from(file.as_raw_fd()), RetrieveOptions::default())?;
    request.connect_response(|r: Response<Basic>| {
        println!("{:#?}", r);
        Ok(())
    })?;

    Ok(())
}

Structs

AsyncSecretProxy
Asynchronous sibling of SecretProxy. The interface lets sandboxed applications retrieve a per-application secret. The secret can then be used for encrypting confidential data inside the sandbox.

RetrieveOptions
Specified options on a retrieve secret request.

SecretProxy
The interface lets sandboxed applications retrieve a per-application secret. The secret can then be used for encrypting confidential data inside the sandbox.