Struct ashpd::documents::FileTransferProxy

pub struct FileTransferProxy<'a>(_);

The interface operates as a middle-man between apps when transferring files via drag-and-drop or copy-paste, taking care of the necessary exporting of files in the document portal.

Toolkits are expected to use the application/vnd.portal.filetransfer mimetype when using this mechanism for file exchange via copy-paste or drag-and-drop.

The data that is transmitted with this mimetype should be the key returned by the StartTransfer method. Upon receiving this mimetype, the target should call RetrieveFiles with the key, to obtain the list of files. The portal will take care of exporting files in the document store as necessary to make them accessible to the target.

Implementations

impl<'a> FileTransferProxy<'a>

pub async fn new(
connection: &Connection
) -> Result<FileTransferProxy<'a>, Error>

Create a new instance of FileTransferProxy.

pub fn inner(&self) -> &Proxy<'_>

Get a reference to the underlying Proxy.

pub async fn add_files(&self, key: &str, fds: &[Fd]) -> Result<(), Error>

Adds files to a session. This method can be called multiple times on a given session. Note that only regular files (not directories) can be added.

Arguments

key - A key returned by FileTransferProxy::start_transfer.

fds - A list of file descriptors of the files to register.

pub async fn retrieve_files(&self, key: &str) -> Result<Vec<String>, Error>

Retrieves files that were previously added to the session with add_files. The files will be exported in the document portal as-needed for the caller, and they will be writable if the owner of the session allowed it.

Returns the list of file paths.

Arguments

key - A key returned by FileTransferProxy::start_transfer.

pub async fn start_transfer(
&self,
writeable: bool,
auto_stop: bool
) -> Result<String, Error>

Starts a session for a file transfer. The caller should call FileTransferProxy::add_files at least once, to add files to this session.

Arguments

writeable - Sets whether the chosen application can write to the files or not.

auto_stop - Whether to stop the transfer automatically after the first FileTransferProxy::retrieve_files call.

Returns

a key that can be passed to FileTransferProxy::retrieve_files to obtain the files.

pub async fn stop_transfer(&self, key: &str) -> Result<(), Error>

Ends the transfer. Further calls to FileTransferProxy::add_files or FileTransferProxy::retrieve_files for this key will return an error.

Arguments

key - A key returned by FileTransferProxy::start_transfer.

pub async fn transfer_closed(&self) -> Result<String, Error>

Emitted when the transfer is closed.

Returns

The key returned by FileTransferProxy::start_transfer