Enum ashpd::WindowIdentifier

pub enum WindowIdentifier {
    Gtk4 {
        root: Root,
        handle: String,
    },
    Gtk3 {
        handle: String,
    },
    Other(String),
}

Most portals interact with the user by showing dialogs. These dialogs should generally be placed on top of the application window that triggered them. To arrange this, the compositor needs to know about the application window. Many portal requests expect a WindowIdentifier for this reason.

Under X11, the WindowIdentifier should have the form x11:XID, where XID is the XID of the application window. Under Wayland, it should have the form wayland:HANDLE, where HANDLE is a surface handle obtained with the xdg-foreign protocol.

For other windowing systems, or if you don’t have a suitable handle, just use the Default implementation.

Please note that the From<gtk3::Window> implementation is x11 only for now.

We would love merge requests that adds other From<T> for WindowIdentifier implementations for other toolkits.

Variants

Gtk4

Gtk 4 Window Identifier

Fields of Gtk4

root: Root

The top level window

handle: String

The exported window handle

Gtk3

GTK 3 Window Identifier

Fields of Gtk3

handle: String

The exported window handle

Other(String)

For Other Toolkits

Implementations

impl WindowIdentifier

pub fn new(identifier: &str) -> Self

Create a new window identifier

impl WindowIdentifier

pub async fn from_window<W: IsA<Root>>(win: &W) -> Self

Creates a WindowIdentifier from a gtk::Root. gtk::Root is the interface implemented by all the widgets that can act as a top level widget.

The constructor returns a valid handle under both Wayland & x11.

Note The function has to be async as the Wayland handle retrieval API is async as well.