Module ashpd::flatpak::update_monitor

Monitor if there's an update it and install it.

Examples

How to monitor if there's a new update and install it. Only available for Flatpak applications.

use ashpd::flatpak::update_monitor::{
    UpdateInfo, UpdateMonitorProxy, UpdateOptions, UpdateProgress,
};
use ashpd::flatpak::{CreateMonitorOptions, FlatpakProxy};
use zbus::{fdo::Result, Connection};
use ashpd::WindowIdentifier;

fn main() -> Result<()> {
    let connection = Connection::new_session()?;
    let proxy = FlatpakProxy::new(&connection)?;

    let monitor_handle = proxy.create_update_monitor(CreateMonitorOptions::default())?;
    let monitor = UpdateMonitorProxy::new(&connection, &monitor_handle)?;

    monitor.on_progress(|p: UpdateProgress| {
        println!("{:#?}", p);
        if p.progress == Some(100) {
            monitor.close().unwrap();
        }
    })?;

    monitor.on_update_available(|_: UpdateInfo| {
        monitor
            .update(WindowIdentifier::default(), UpdateOptions::default())
            .unwrap();
    })?;

    Ok(())
}

Structs

UpdateInfo
A response containing the update information when an update is available.

UpdateMonitorProxy
The interface exposes some interactions with Flatpak on the host to the sandbox. For example, it allows you to restart the applications or start a more sandboxed instance.

UpdateOptions
Specficied options on an update request

UpdateProgress
A response of the update progress signal.

Enums

UpdateStatus
The update status.