ashpd
Module desktop

Interact with the user’s desktop such as taking a screenshot, setting a background or querying the user’s location.

Modules

account

Access to the current logged user information such as the id, name or their avatar uri.

background

Request to run in the background or started automatically when the user logs in.

camera

Check if a camera is available, request access to it and open a PipeWire remote stream.

clipboard

Interact with the clipboard.

deviceDeprecated

Request access to specific devices such as camera, speakers or microphone.

dynamic_launcher

Install launchers like Web Application from your browser or Steam.

email

Compose an email.

file_chooser

Open/save file(s) chooser. The interface lets sandboxed applications ask the user for access to files outside the sandbox. The portal backend will present the user with a file chooser dialog.

game_mode

Enable/disable/query the status of Game Mode.

global_shortcuts

Register global shortcuts Register global shortcuts

inhibit

Inhibit the session from being restarted or the user from logging out.

input_capture

Capture input events from physical or logical devices.

location

Query the user’s GPS location.

memory_monitor

Monitor memory level.

network_monitor

Check the status of the network on a user’s machine. Note This portal doesn’t work for sandboxed applications.

notification

Send/withdraw notifications.

open_uri

Open a URI or a directory.

power_profile_monitor

Power profile monitoring.

print

Print a document.

proxy_resolver

Proxy information. Note This portal doesn’t work for sandboxed applications.

realtime

Set threads to realtime.

remote_desktop

Start a remote desktop session and interact with it.

screencast

Start a screencast session and get the PipeWire remote of it.

screenshot

Take a screenshot or pick a color.

secret

Retrieve a per-application secret used to encrypt confidential data inside the sandbox.

settings

Read & listen to system settings changes.

trash

Move a file to the trash.

wallpaper

Set a wallpaper on lockscreen, background or both.

Structs

Color

A color as a RGB tuple.

Request

The Request interface is shared by all portal interfaces.

Session

Shared by all portal interfaces that involve long lived sessions.

Enums

Icon

A representation of an icon.

PersistMode

Persistence mode for a screencast or remote desktop session.

Response

A typical response returned by the Request::response. of a Request.

ResponseError

An error returned a portal request caused by either the user cancelling the request or something else.

ResponseType

Possible responses.

Traits

SessionPortal

Portals that have a long-lived interaction