Module ashpd::desktop::account

Request access to the current logged user information such as the id, name or their avatar uri.

Examples

use ashpd::desktop::account::{AccountProxy, UserInfoOptions, UserInfo};
use ashpd::{RequestProxy, Response, WindowIdentifier};
use zbus::fdo::Result;

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = AccountProxy::new(&connection)?;
    let request_handle = proxy.get_user_information(
        WindowIdentifier::default(),
        UserInfoOptions::default()
            .reason("Fractal would like access to your information"),
    )?;
    let req = RequestProxy::new(&connection, &request_handle)?;
    req.on_response(|response: Response<UserInfo>| {
        if let Ok(user_info) = response {
            println!("{}", user_info.id);
            println!("{}", user_info.name);
            println!("{}", user_info.image);
        }
    })?;
    Ok(())
}

Structs

AccountProxy
The interface lets sandboxed applications query basic information about the user, like his name and avatar photo.

UserInfo
The response of a get_user_information request.

UserInfoOptions
The possible options for a get user information request.