Module ashpd::desktop::background

Request running an application in the background.

Examples

use ashpd::desktop::background::{
    BackgroundOptions, BackgroundProxy, Background,
};
use ashpd::{RequestProxy, Response, WindowIdentifier};
use zbus::fdo::Result;

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = BackgroundProxy::new(&connection)?;

    let request_handle = proxy.request_background(
        WindowIdentifier::default(),
        BackgroundOptions::default()
            .autostart(true)
            .commandline(vec!["geary".to_string()])
            .reason("Automatically fetch your latest mails."),
    )?;

    let request = RequestProxy::new(&connection, &request_handle)?;
    request.on_response(|response: Response<Background>| {
        if let Ok(bg) = response {
            println!("{}", bg.autostart);
            println!("{}", bg.background);
        }
    })?;
    Ok(())
}

Structs

Background
The response of a background request.

BackgroundOptions
Specified options for a request_background request.

BackgroundProxy
The interface lets sandboxed applications request that the application is allowed to run in the background or started automatically when the user logs in.