ashpd::desktop
Module device

👎Deprecated: The portal does not serve any purpose as nothing really can make use of it as is.

Request access to specific devices such as camera, speakers or microphone.

Note This portal doesn’t work for sandboxed applications.

Examples

Access a Device

use ashpd::desktop::device::{Device, DeviceProxy};

async fn run() -> ashpd::Result<()> {
    let proxy = DeviceProxy::new().await?;
    proxy.access_device(6879, &[Device::Speakers]).await?;
    Ok(())
}

Structs

DeviceProxyDeprecated

The interface lets services ask if an application should get access to devices such as microphones, speakers or cameras.

Enums

DeviceDeprecated

The possible device to request access to.