ashpd::desktop
Module global_shortcuts

Register global shortcuts Register global shortcuts

Structs

Activated

Notifies about a shortcut becoming active.

BindShortcuts

A response to a GlobalShortcuts::bind_shortcuts request.

Deactivated

Notifies that a shortcut is not active anymore.

GlobalShortcuts

Wrapper of the DBus interface: org.freedesktop.portal.GlobalShortcuts.

ListShortcuts

A response to a GlobalShortcuts::list_shortcuts request.

NewShortcut

Shortcut descriptor used to bind new shortcuts in GlobalShortcuts::bind_shortcuts

Shortcut

Struct that contains information about existing binded shortcut.

ShortcutsChanged

Indicates that the information associated with some of the shortcuts has changed.