Module ashpd::desktop::network_monitor

Check the status of the network on a user's machine.

Examples

use ashpd::desktop::network_monitor::NetworkMonitorProxy;
use zbus::fdo::Result;

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = NetworkMonitorProxy::new(&connection)?;

    println!("{}", proxy.can_reach("www.google.com", 80)?);

    println!("{}", proxy.get_available()?);

    println!("{:#?}", proxy.get_connectivity()?);

    println!("{}", proxy.get_metered()?);

    println!("{:#?}", proxy.get_status()?);

    Ok(())
}

Structs

NetworkMonitorProxy
The interface provides network status information to sandboxed applications. It is not a portal in the strict sense, since it does not involve user interaction. Applications are expected to use this interface indirectly, via a library API such as the GLib GNetworkMonitor interface.

NetworkStatus
The network status, composed of the avaiability, metered & connectivity

Enums

Connectivity
Host's network activity