Struct ashpd::desktop::print::Settings

pub struct Settings {

    pub orientation: Option<Orientation>,
    pub paper_format: Option<String>,
    pub paper_width: Option<String>,
    pub paper_height: Option<String>,
    pub n_copies: Option<String>,
    pub default_source: Option<String>,
    pub quality: Option<Quality>,
    pub resolution: Option<String>,
    pub use_color: Option<bool>,
    pub duplex: Option<String>,
    pub collate: Option<String>,
    pub reverse: Option<String>,
    pub media_type: Option<String>,
    pub dither: Option<String>,
    pub scale: Option<String>,
    pub print_pages: Option<String>,
    pub page_ranges: Option<String>,
    pub page_set: Option<String>,
    pub finishings: Option<String>,
    pub number_up: Option<String>,
    pub number_up_layout: Option<String>,
    pub output_bin: Option<String>,
    pub resolution_x: Option<String>,
    pub resolution_y: Option<String>,
    pub print_lpi: Option<String>,
    pub output_basename: Option<String>,
    pub output_file_format: Option<String>,
    pub output_uri: Option<Url>,

}

Print settings to set in the print dialog.

Fields
orientation: Option<Orientation>

One of landscape, portrait, reverse_landscape or reverse_portrait.

paper_format: Option<String>

A paper name according to PWG 5101.1-2002

paper_width: Option<String>

Paper width, in millimeters.

paper_height: Option<String>

Paper height, in millimeters.

n_copies: Option<String>

The number of copies to print.

default_source: Option<String>

The default paper source.

quality: Option<Quality>

Print quality.

resolution: Option<String>

The resolution, sets both resolution-x & resolution-y

use_color: Option<bool>

Whether to use color.

duplex: Option<String>

Duplex printing mode, one of simplex, horizontal or vertical.

collate: Option<String>

Whether to collate copies.

reverse: Option<String>

Whether to reverse the order of printed pages.

media_type: Option<String>

A media type according to PWG 5101.1-2002

dither: Option<String>

The dithering to use, one of fine, none, coarse, lineart, grayscale or error-diffusion.

scale: Option<String>

The scale in percent

print_pages: Option<String>

What pages to print, one of all, selection, current or ranges.

page_ranges: Option<String>

A list of page ranges, formatted like this: 0-2,4,9-11.

page_set: Option<String>

What pages to print, one of all, even or odd.

finishings: Option<String>

The finishings.

number_up: Option<String>

The number of pages per sheet.

number_up_layout: Option<String>

One of lrtb, lrbt, rltb, rlbt, tblr, tbrl, btlr, btrl.

output_bin: Option<String>

The output bin.

resolution_x: Option<String>

The horizontal resolution in dpi.

resolution_y: Option<String>

The vertical resolution in dpi.

print_lpi: Option<String>

The resolution in lpi (lines per inch).

output_basename: Option<String>

Basename to use for print-to-file.

output_file_format: Option<String>

Format to use for print-to-file, one of PDF, PS, SVG

output_uri: Option<Url>

The uri used for print-to file.

Implementations

impl Settings

pub fn orientation(self, orientation: Orientation) -> Self

Sets the orientation.

pub fn paper_format(self, paper_format: &str) -> Self

Sets the paper name.

pub fn paper_width(self, paper_width: &str) -> Self

Sets the paper width.

pub fn paper_height(self, paper_height: &str) -> Self

Sets the paper height.

pub fn n_copies(self, n_copies: &str) -> Self

Sets the number of copies to print.

pub fn default_source(self, default_source: &str) -> Self

Sets the default paper source.

pub fn quality(self, quality: Quality) -> Self

Sets the print quality.

pub fn resolution(self, resolution: &str) -> Self

Sets the resolution, both resolution-x & resolution-y.

pub fn use_color(self, use_color: bool) -> Self

Sets whether to use color.

pub fn duplex(self, duplex: &str) -> Self

Sets the duplex printing mode.

pub fn collate(self, collate: &str) -> Self

Whether to collate copies.

pub fn reverse(self, reverse: &str) -> Self

Sets whether to reverse the order of the printed pages.

pub fn media_type(self, media_type: &str) -> Self

Sets the media type.

pub fn dither(self, dither: &str) -> Self

Sets the dithering to use.

pub fn scale(self, scale: &str) -> Self

Sets the page scale in percent.

pub fn print_pages(self, print_pages: &str) -> Self

Sets what pages to print, one of all, selection, current or ranges.

pub fn page_ranges(self, page_ranges: &str) -> Self

Sets a list of page ranges, formatted like this: 0-2,4,9-11.

pub fn page_set(self, page_set: &str) -> Self

Sets what pages to print, one of all, even or odd.

pub fn finishings(self, finishings: &str) -> Self

Sets the finishings.

pub fn number_up(self, number_up: &str) -> Self

Sets the number of pages per sheet.

pub fn number_up_layout(self, number_up_layout: &str) -> Self

Sets the number up layout, one of lrtb, lrbt, rltb, rlbt, tblr, tbrl, btlr, btrl.

pub fn output_bin(self, output_bin: &str) -> Self

Sets the output bin

pub fn resolution_x(self, resolution_x: &str) -> Self

Sets the horizontal resolution in dpi.

pub fn resolution_y(self, resolution_y: &str) -> Self

Sets the vertical resolution in dpi.

pub fn print_lpi(self, print_lpi: &str) -> Self

Sets the resolution in lines per inch.

pub fn output_basename(self, output_basename: &str) -> Self

Sets the print-to-file base name.

pub fn output_file_format(self, output_file_format: &str) -> Self

Sets the print-to-file format, one of PS, PDF, SVG.

pub fn output_uri(self, output_uri: &Url) -> Self

Sets the print-to-file output uri.