Module ashpd::desktop::remote_desktop

Start a remote desktop session and interact with it.

Examples

use ashpd::desktop::remote_desktop::{
    CreateRemoteOptions, CreateSession, DeviceType, KeyState, RemoteDesktopProxy, SelectDevices,
    SelectDevicesOptions, StartRemoteOptions,
};
use ashpd::{BasicResponse as Basic, HandleToken, RequestProxy, Response, WindowIdentifier};
use zbus::{fdo::Result, Connection};
use zvariant::ObjectPath;
use std::collections::HashMap;
use std::convert::TryFrom;

fn select_devices(
    handle: ObjectPath,
    connection: &Connection,
    proxy: &RemoteDesktopProxy,
) -> Result<()> {
    let request_handle = proxy.select_devices(
        handle.clone(),
        SelectDevicesOptions::default().types(DeviceType::Keyboard | DeviceType::Pointer),
    )?;

    let request = RequestProxy::new(&connection, &request_handle)?;
    request.on_response(|r: Response<Basic>| {
        if r.is_ok() {
            start_remote(handle, connection, proxy).unwrap();
        }
    })?;

    Ok(())
}

fn start_remote(
    handle: ObjectPath,
    connection: &Connection,
    proxy: &RemoteDesktopProxy,
) -> Result<()> {
    let request_handle = proxy.start(
        handle.clone(),
        WindowIdentifier::default(),
        StartRemoteOptions::default(),
    )?;

    let request = RequestProxy::new(&connection, &request_handle)?;
    request.on_response(|r: Response<SelectDevices>| {
        proxy
            .notify_keyboard_keycode(
                handle.clone(),
                HashMap::new(),
                13, // Enter key code
                KeyState::Pressed,
            )
            .unwrap();

        println!("{:#?}", r.unwrap().devices);
    })?;

    Ok(())
}

fn main() -> Result<()> {
    let connection = Connection::new_session()?;
    let proxy = RemoteDesktopProxy::new(&connection)?;

    let handle = proxy.create_session(
        CreateRemoteOptions::default()
            .session_handle_token(HandleToken::try_from("token").unwrap()),
    )?;

    let request = RequestProxy::new(&connection, &handle)?;
    request.on_response(|r: Response<CreateSession>| {
        let session = r.unwrap();
        select_devices(session.handle(), &connection, &proxy).unwrap();
    })?;

    Ok(())
}

Structs

CreateRemoteOptions
Specified options on a create a remote session request.

CreateSession

RemoteDesktopProxy
The interface lets sandboxed applications create remote desktop sessions.

SelectDevices

SelectDevicesOptions
Specified options on a select devices request.

StartRemoteOptions
Specified options on a start remote desktop request.

Enums

Axis

DeviceType

KeyState