Module ashpd::desktop::remote_desktop

Start a remote desktop session and interact with it.

Examples

use ashpd::desktop::remote_desktop::{DeviceType, KeyState, RemoteDesktopProxy};
use ashpd::WindowIdentifier;

async fn run() -> ashpd::Result<()> {
    let connection = zbus::Connection::session().await?;
    let proxy = RemoteDesktopProxy::new(&connection).await?;

    let session = proxy.create_session().await?;

    proxy.select_devices(&session, DeviceType::Keyboard | DeviceType::Pointer).await?;

    let (devices, _) = proxy.start(&session, &WindowIdentifier::default()).await?;
    println!("{:#?}", devices);

    // 13 for Enter key code
    proxy.notify_keyboard_keycode(&session, 13, KeyState::Pressed).await?;

    Ok(())
}

You can also use the Remote Desktop portal with the ScreenCast one. In order to do so, you need to call ScreenCastProxy::select_sources() on the session created with RemoteDesktopProxy::create_session()

use ashpd::desktop::remote_desktop::{DeviceType, KeyState, RemoteDesktopProxy};
use ashpd::desktop::screencast::{CursorMode, ScreenCastProxy, SourceType};
use ashpd::WindowIdentifier;

async fn run() -> ashpd::Result<()> {
    let connection = zbus::Connection::session().await?;
    let proxy = RemoteDesktopProxy::new(&connection).await?;
    let screencast = ScreenCastProxy::new(&connection).await?;
    let identifier = WindowIdentifier::default();

    let session = proxy.create_session().await?;

    proxy.select_devices(&session, DeviceType::Keyboard | DeviceType::Pointer).await?;
    screencast
        .select_sources(
            &session,
            CursorMode::Metadata.into(),
            SourceType::Monitor | SourceType::Window,
            true,
        )
        .await?;

    let (devices, streams) = proxy.start(&session, &identifier).await?;
    println!("{:#?}", devices);
    println!("{:#?}", streams);

    // 13 for Enter key code
    proxy.notify_keyboard_keycode(&session, 13, KeyState::Pressed).await?;

    Ok(())
}

Structs

RemoteDesktopProxy

The interface lets sandboxed applications create remote desktop sessions.

Enums

Axis

The available axis.

DeviceType

A bit flag for the available devices.

KeyState

The keyboard key state.