Struct ashpd::desktop::remote_desktop::RemoteDesktopProxy

pub struct RemoteDesktopProxy<'c>(_);

The interface lets sandboxed applications create remote desktop sessions.

Implementations

impl<'c> RemoteDesktopProxy<'c>

pub fn new(conn: &Connection) -> Result<Self>

Creates a new proxy with the default service & path.

pub fn new_for(
conn: &Connection,
destination: &'c str,
path: &'c str
) -> Result<Self>

Creates a new proxy for the given destination and path.

pub fn new_for_owned(
conn: Connection,
destination: String,
path: String
) -> Result<Self>

Same as new_for but takes ownership of the passed arguments.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'_>

The reference to the underlying zbus::Proxy.

pub fn create_session(
&self,
options: CreateRemoteOptions
) -> Result<OwnedObjectPath>

Create a remote desktop session. A remote desktop session is used to allow remote controlling a desktop session. It can also be used together with a screen cast session

Returns a RequestProxy object path.

Arguments

options - A CreateRemoteOptions

pub fn notify_keyboard_keycode(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
keycode: i32,
state: KeyState
) -> Result<()>

Notify keyboard code May only be called if KEYBOARD access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

keycode - Keyboard code that was pressed or released

state - The new state of the keyboard code

FIXME: figure out the options we can take here

pub fn notify_keyboard_keysym(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
keysym: i32,
state: KeyState
) -> Result<()>

Notify keyboard symbol May only be called if KEYBOARD access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

keysym - Keyboard symbol that was pressed or released

state - The new state of the keyboard code

FIXME: figure out the options we can take here

pub fn notify_pointer_axis(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
dx: f64,
dy: f64
) -> Result<()>

Notify pointer axis The axis movement from a 'smooth scroll' device, such as a touchpad. When applicable, the size of the motion delta should be equivalent to the motion vector of a pointer motion done using the same advice.

May only be called if POINTER access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

dx - Relative axis movement on the x axis

dy - Relative axis movement on the y axis

FIXME: figure out the options we can take here

pub fn notify_pointer_axis_discrete(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
axis: Axis,
steps: i32
) -> Result<()>

Notify pointer axis discrete May only be called if POINTER access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

axis - The axis that was scrolled

FIXME: figure out the options we can take here

pub fn notify_pointer_button(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
button: i32,
state: KeyState
) -> Result<()>

Notify pointer button The pointer button is encoded according to Linux Evdev button codes.

May only be called if POINTER access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

button - The pointer button was pressed or released

state - The new state of the keyboard code

FIXME: figure out the options we can take here

pub fn notify_pointer_motion(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
dx: f64,
dy: f64
) -> Result<()>

Notify about a new relative pointer motion event. The (dx, dy) vector represents the new pointer position in the streams logical coordinate space.

Arguments

session_handle - A SessionProxy object path.

options - ?

dx - Relative movement on the x axis

dy - Relative movement on the y axis

FIXME: figure out the options we can take here

pub fn notify_pointer_motion_absolute(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
stream: u32,
x: f64,
y: f64
) -> Result<()>

Notify about a new absolute pointer motion event. The (x, y) position represents the new pointer position in the streams logical coordinate sspace

Arguments

session_handle - A SessionProxy object path.

options - ?

stream - The PipeWire stream node the coordinate is relative to

x - Pointer motion x coordinate

y - Pointer motion y coordinate

FIXME: figure out the options we can take here

pub fn notify_touch_down(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<()>

Notify about a new touch down event. The (x, y) position represents the new touch point position in the streams logical coordinate space

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

stream - The PipeWire stream node the coordinate is relative to

slot - Touch slot where touch point appeared

x - Touch down x coordinate

y - Touch down y coordinate

FIXME: figure out the options we can take here

pub fn notify_touch_motion(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<()>

Notify about a new touch motion event. The (x, y) position represents where the touch point position in the streams logical coordinate space moved

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

stream - The PipeWire stream node the coordinate is relative to

slot - Touch slot where touch point appeared

x - Touch motion x coordinate

y - Touch motion y coordinate

FIXME: figure out the options we can take here

pub fn notify_touch_up(
&self,
session_handle: ObjectPath<'_>,
options: HashMap<&str, Value<'_>>,
slot: u32
) -> Result<()>

Notify about a new touch up event.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session_handle - A SessionProxy object path.

options - ?

slot - Touch slot where touch point appeared

FIXME: figure out the options we can take here

pub fn select_devices(
&self,
session_handle: ObjectPath<'_>,
options: SelectDevicesOptions
) -> Result<OwnedObjectPath>

Select input devices to remote control.

Returns a RequestProxy object path.

Arguments

session_handle - A SessionProxy object path.

options - SelectDevicesOptions

pub fn start(
&self,
session_handle: ObjectPath<'_>,
parent_window: WindowIdentifier,
options: StartRemoteOptions
) -> Result<OwnedObjectPath>

Start the remote desktop session.

This will typically result in the portal presenting a dialog letting the user select what to share, including devices and optionally screen content if screen cast sources was selected.

Returns a RequestProxy object path.

Arguments

session_handle - A SessionProxy object path.

parent_window - The application window identifier

options - StartRemoteOptions

pub fn available_device_types(&self) -> Result<u32>

Available source types.

pub fn version(&self) -> Result<u32>

version property

pub fn introspect(&self) -> Result<String>