Struct ashpd::desktop::remote_desktop::RemoteDesktopProxy

pub struct RemoteDesktopProxy<'a>(_);

The interface lets sandboxed applications create remote desktop sessions.

Wrapper of the DBus interface: org.freedesktop.portal.RemoteDesktop.

Implementations

impl<'a> RemoteDesktopProxy<'a>

pub async fn new() -> Result<RemoteDesktopProxy<'a>, Error>

Create a new instance of RemoteDesktopProxy.

pub fn inner(&self) -> &Proxy<'_>

Get a reference to the underlying Proxy.

pub async fn create_session(&self) -> Result<SessionProxy<'a>, Error>

Create a remote desktop session. A remote desktop session is used to allow remote controlling a desktop session. It can also be used together with a screen cast session.

Specifications

See also CreateSession.

pub async fn select_devices(
&self,
session: &SessionProxy<'_>,
types: BitFlags<DeviceType>
) -> Result<(), Error>

Select input devices to remote control.

Arguments

session - A SessionProxy, created with create_session().

types - The device types to request remote controlling of.

Specifications

See also SelectDevices.

pub async fn start(
&self,
session: &SessionProxy<'_>,
identifier: &WindowIdentifier
) -> Result<(BitFlags<DeviceType>, Vec<Stream>), Error>

Start the remote desktop session.

This will typically result in the portal presenting a dialog letting the user select what to share, including devices and optionally screen content if screen cast sources was selected.

Arguments

session - A SessionProxy, created with create_session().

identifier - The application window identifier.

Specifications

See also Start.

pub async fn notify_keyboard_keycode(
&self,
session: &SessionProxy<'_>,
keycode: i32,
state: KeyState
) -> Result<(), Error>

Notify keyboard code.

Note only works if DeviceType::Keyboard access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

keycode - Keyboard code that was pressed or released.

state - The new state of the keyboard code.

Specifications

See also NotifyKeyboardKeycode.

pub async fn notify_keyboard_keysym(
&self,
session: &SessionProxy<'_>,
keysym: i32,
state: KeyState
) -> Result<(), Error>

Notify keyboard symbol.

Note only works if DeviceType::Keyboard access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

keysym - Keyboard symbol that was pressed or released.

state - The new state of the keyboard code.

Specifications

See also NotifyKeyboardKeysym.

pub async fn notify_touch_up(
&self,
session: &SessionProxy<'_>,
slot: u32
) -> Result<(), Error>

Notify about a new touch up event.

Note only works if DeviceType::Touchscreen access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

slot - Touch slot where touch point appeared.

Specifications

See also NotifyTouchUp.

pub async fn notify_touch_down(
&self,
session: &SessionProxy<'_>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new touch down event. The (x, y) position represents the new touch point position in the streams logical coordinate space.

Note only works if DeviceType::Touchscreen access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch down x coordinate.

y - Touch down y coordinate.

Specifications

See also NotifyTouchDown.

pub async fn notify_touch_motion(
&self,
session: &SessionProxy<'_>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new touch motion event. The (x, y) position represents where the touch point position in the streams logical coordinate space moved.

Note only works if DeviceType::Touchscreen access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch motion x coordinate.

y - Touch motion y coordinate.

Specifications

See also NotifyTouchMotion.

pub async fn notify_pointer_motion_absolute(
&self,
session: &SessionProxy<'_>,
stream: u32,
x: f64,
y: f64
) -> Result<(), Error>

Notify about a new absolute pointer motion event. The (x, y) position represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy, created with create_session().

stream - The PipeWire stream node the coordinate is relative to.

x - Pointer motion x coordinate.

y - Pointer motion y coordinate.

Specifications

See also NotifyPointerMotionAbsolute.

pub async fn notify_pointer_motion(
&self,
session: &SessionProxy<'_>,
dx: f64,
dy: f64
) -> Result<(), Error>

Notify about a new relative pointer motion event. The (dx, dy) vector represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy, created with create_session().

dx - Relative movement on the x axis.

dy - Relative movement on the y axis.

Specifications

See also NotifyPointerMotion.

pub async fn notify_pointer_button(
&self,
session: &SessionProxy<'_>,
button: i32,
state: KeyState
) -> Result<(), Error>

Notify pointer button. The pointer button is encoded according to Linux Evdev button codes.

Note only works if DeviceType::Pointer access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

button - The pointer button was pressed or released.

state - The new state of the keyboard code.

Specifications

See also NotifyPointerButton.

pub async fn notify_pointer_axis_discrete(
&self,
session: &SessionProxy<'_>,
axis: Axis,
steps: i32
) -> Result<(), Error>

Notify pointer axis discrete.

Note only works if DeviceType::Pointer access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

axis - The axis that was scrolled.

Specifications

See also NotifyPointerAxisDiscrete.

pub async fn notify_pointer_axis(
&self,
session: &SessionProxy<'_>,
dx: f64,
dy: f64,
finish: bool
) -> Result<(), Error>

Notify pointer axis. The axis movement from a “smooth scroll” device, such as a touchpad. When applicable, the size of the motion delta should be equivalent to the motion vector of a pointer motion done using the same advice.

Note only works if DeviceType::Pointer access was provided after starting the session.

Arguments

session - A SessionProxy, created with create_session().

dx - Relative axis movement on the x axis.

dy - Relative axis movement on the y axis.

finish - Whether it is the last axis event.

Specifications

See also NotifyPointerAxis.

pub async fn available_device_types(
&self
) -> Result<BitFlags<DeviceType>, Error>

Available source types.

Specifications

See also AvailableDeviceTypes.