Struct ashpd::desktop::remote_desktop::RemoteDesktopProxy

pub struct RemoteDesktopProxy<'c>(_);

The interface lets sandboxed applications create remote desktop sessions.

Implementations

impl<'c> RemoteDesktopProxy<'c>

pub fn new(conn: &Connection) -> Result<Self>

Creates a new proxy with the default service & path.

pub fn new_owned(conn: Connection) -> Result<Self>

Creates a new proxy with the default service & path, taking ownership of conn.

pub fn new_for<E>(
conn: &Connection,
destination: &'c str,
path: impl TryInto<ObjectPath<'c>, Error = E>
) -> Result<Self> where
Error: From<E>,

Creates a new proxy for the given destination and path.

pub fn new_for_owned<E>(
conn: Connection,
destination: String,
path: impl TryInto<ObjectPath<'static>, Error = E>
) -> Result<Self> where
Error: From<E>,

Same as new_for but takes ownership of the passed arguments.

pub fn new_for_path<E>(
conn: &Connection,
path: impl TryInto<ObjectPath<'c>, Error = E>
) -> Result<Self> where
Error: From<E>,

Creates a new proxy for the given path.

pub fn new_for_owned_path<E>(
conn: Connection,
path: impl TryInto<ObjectPath<'static>, Error = E>
) -> Result<Self> where
Error: From<E>,

Same as new_for_path but takes ownership of the passed arguments.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'_>

The reference to the underlying zbus::Proxy.

pub fn create_session(
&self,
options: CreateRemoteOptions
) -> Result<RequestProxy<'_>>

Create a remote desktop session. A remote desktop session is used to allow remote controlling a desktop session. It can also be used together with a screen cast session.

Arguments

options - A CreateRemoteOptions.

pub fn notify_keyboard_keycode<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
keycode: i32,
state: KeyState
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify keyboard code. May only be called if KEYBOARD access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

keycode - Keyboard code that was pressed or released.

state - The new state of the keyboard code.

FIXME: figure out the options we can take here

pub fn notify_keyboard_keysym<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
keysym: i32,
state: KeyState
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify keyboard symbol. May only be called if KEYBOARD access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

keysym - Keyboard symbol that was pressed or released.

state - The new state of the keyboard code.

FIXME: figure out the options we can take here

pub fn notify_pointer_axis<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
dx: f64,
dy: f64
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify pointer axis. The axis movement from a "smooth scroll" device, such as a touchpad. When applicable, the size of the motion delta should be equivalent to the motion vector of a pointer motion done using the same advice.

May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

dx - Relative axis movement on the x axis.

dy - Relative axis movement on the y axis.

FIXME: figure out the options we can take here

pub fn notify_pointer_axis_discrete<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
axis: Axis,
steps: i32
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify pointer axis discrete. May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

axis - The axis that was scrolled.

FIXME: figure out the options we can take here

pub fn notify_pointer_button<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
button: i32,
state: KeyState
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify pointer button. The pointer button is encoded according to Linux Evdev button codes.

May only be called if POINTER access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

button - The pointer button was pressed or released.

state - The new state of the keyboard code.

FIXME: figure out the options we can take here

pub fn notify_pointer_motion<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
dx: f64,
dy: f64
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify about a new relative pointer motion event. The (dx, dy) vector represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

dx - Relative movement on the x axis.

dy - Relative movement on the y axis.

FIXME: figure out the options we can take here

pub fn notify_pointer_motion_absolute<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
stream: u32,
x: f64,
y: f64
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify about a new absolute pointer motion event. The (x, y) position represents the new pointer position in the streams logical coordinate space.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

stream - The PipeWire stream node the coordinate is relative to.

x - Pointer motion x coordinate.

y - Pointer motion y coordinate.

FIXME: figure out the options we can take here

pub fn notify_touch_down<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify about a new touch down event. The (x, y) position represents the new touch point position in the streams logical coordinate space.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch down x coordinate.

y - Touch down y coordinate.

FIXME: figure out the options we can take here

pub fn notify_touch_motion<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
stream: u32,
slot: u32,
x: f64,
y: f64
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify about a new touch motion event. The (x, y) position represents where the touch point position in the streams logical coordinate space moved.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

stream - The PipeWire stream node the coordinate is relative to.

slot - Touch slot where touch point appeared.

x - Touch motion x coordinate.

y - Touch motion y coordinate.

FIXME: figure out the options we can take here

pub fn notify_touch_up<S>(
&self,
session: &S,
options: HashMap<&str, Value<'_>>,
slot: u32
) -> Result<()> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Notify about a new touch up event.

May only be called if TOUCHSCREEN access was provided after starting the session.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - ?

slot - Touch slot where touch point appeared.

FIXME: figure out the options we can take here

pub fn select_devices<S>(
&self,
session: &S,
options: SelectDevicesOptions
) -> Result<RequestProxy<'_>> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Select input devices to remote control.

Arguments

session - A SessionProxy or AsyncSessionProxy.

options - SelectDevicesOptions.

pub fn start<S>(
&self,
session: &S,
parent_window: WindowIdentifier,
options: StartRemoteOptions
) -> Result<RequestProxy<'_>> where
S: Into<SessionProxy<'c>> + Serialize + Type,

Start the remote desktop session.

This will typically result in the portal presenting a dialog letting the user select what to share, including devices and optionally screen content if screen cast sources was selected.

Arguments

session - A SessionProxy or AsyncSessionProxy.

parent_window - The application window identifier.

options - StartRemoteOptions.

pub fn available_device_types(&self) -> Result<BitFlags<DeviceType>>

Available source types.

pub fn version(&self) -> Result<u32>

The version of this DBus interface.

pub fn introspect(&self) -> Result<String>

Introspect the associated object, and return the XML description.