Module ashpd::desktop::screencast

Start a screencast session and get the PipeWire remote of it.

Examples

How to create a screen cast session & start it. The portal is currently useless without PipeWire & Rust support.

use ashpd::{
    desktop::screencast::{CursorMode, PersistMode, Screencast, SourceType},
    WindowIdentifier,
};

async fn run() -> ashpd::Result<()> {
    let proxy = Screencast::new().await?;
    let session = proxy.create_session().await?;
    proxy
        .select_sources(
            &session,
            CursorMode::Metadata,
            SourceType::Monitor | SourceType::Window,
            true,
            None,
            PersistMode::DoNot,
        )
        .await?;

    let (streams, token) = proxy.start(&session, &WindowIdentifier::default()).await?;

    streams.iter().for_each(|stream| {
        println!("node id: {}", stream.pipe_wire_node_id());
        println!("size: {:?}", stream.size());
        println!("position: {:?}", stream.position());
    });
    Ok(())
}

Structs

Screencast

The interface lets sandboxed applications create screen cast sessions.

Stream

A PipeWire stream.

Enums

CursorMode

A bit flag for the possible cursor modes.

PersistMode

SourceType

A bit flag for the available sources to record.