Module ashpd::desktop::screencast

Start a screencast session and get the PipeWire remote of it.

Examples

How to create a screen cast session & start it. The portal is currently useless without PipeWire & Rust support.

use ashpd::desktop::screencast::{
    CreateSession, CreateSessionOptions, CursorMode, ScreenCastProxy, SelectSourcesOptions,
    SourceType, StartCastOptions, Streams,
};
use ashpd::{BasicResponse as Basic, HandleToken, Response, SessionProxy, WindowIdentifier};
use enumflags2::BitFlags;
use std::convert::TryFrom;
use zbus::{self, fdo::Result};
use zvariant::ObjectPath;

fn select_sources(
    session: &SessionProxy<'c>,
    proxy: &ScreenCastProxy<'c>,
) -> Result<()> {
    let request = proxy.select_sources(
        session,
        SelectSourcesOptions::default()
            .multiple(true)
            .cursor_mode(BitFlags::from(CursorMode::Metadata))
            .types(SourceType::Monitor | SourceType::Window),
    )?;

    request.connect_response(move |response: Response<Basic>| {
        if response.is_ok() {
            start_cast(session, proxy)?;
        }
        Ok(())
    })?;
    Ok(())
}

fn start_cast(
    session: &SessionProxy<'c>,
    proxy: &ScreenCastProxy<'c>,
) -> Result<()> {
    let request = proxy.start(
        session_handle,
        WindowIdentifier::default(),
        StartCastOptions::default(),
    )?;
    request.connect_response(move |r: Response<Streams>| {
        r.unwrap().streams().iter().for_each(|stream| {
            println!("{}", stream.pipewire_node_id());
            println!("{:#?}", stream.properties());
        });
        Ok(())
    })?;
    Ok(())
}

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = ScreenCastProxy::new(&connection);

    let session_token = HandleToken::try_from("session120").unwrap();

    let request = proxy
        .create_session(CreateSessionOptions::default().session_handle_token(session_token))?;

    request.connect_response(|response: Response<CreateSession>| {
        let response = response.unwrap();
        let session = SessionProxy::new_for_path(&connection, response.session_handle())?;
        select_sources(&session, &proxy)?;
        Ok(())
    })?;
    Ok(())
}

Structs

AsyncScreenCastProxy
Asynchronous sibling of ScreenCastProxy. The interface lets sandboxed applications create screen cast sessions.

CreateSession
A response to the create session request.

CreateSessionOptions
Specified options on a create a Screen Cast session request.

ScreenCastProxy
The interface lets sandboxed applications create screen cast sessions.

SelectSourcesOptions
Specified options on a select sources request.

StartCastOptions
Specified options on a start Screen Cast request.

Stream
A PipeWire stream.

StreamProperties
The stream properties.

Streams
A response to start the Stream Cast request.

Enums

CursorMode
A bit flag for the possible cursor modes.

SourceType
A bit flag for the available sources to record.