Struct ashpd::desktop::screencast::ScreenCastProxy

pub struct ScreenCastProxy<'a>(_);

The interface lets sandboxed applications create screen cast sessions.

Wrapper of the DBus interface: org.freedesktop.portal.ScreenCast.

Implementations

impl<'a> ScreenCastProxy<'a>

pub async fn new(connection: &Connection) -> Result<ScreenCastProxy<'a>, Error>

Create a new instance of ScreenCastProxy.

pub fn inner(&self) -> &Proxy<'_>

Get a reference to the underlying Proxy.

pub async fn create_session(&self) -> Result<SessionProxy<'a>, Error>

Create a screen cast session.

pub async fn open_pipe_wire_remote(
&self,
session: &SessionProxy<'_>
) -> Result<Fd, Error>

Open a file descriptor to the PipeWire remote where the screen cast streams are available.

Returns a file descriptor of an open PipeWire remote.

Arguments

session - A SessionProxy.

pub async fn select_sources(
&self,
session: &SessionProxy<'_>,
cursor_mode: BitFlags<CursorMode>,
types: BitFlags<SourceType>,
multiple: bool
) -> Result<(), Error>

Configure what the screen cast session should record. This method must be called before starting the session.

Passing invalid input to this method will cause the session to be closed. An application may only attempt to select sources once per session.

Arguments

session - A SessionProxy.

cursor_mode - Sets how the cursor will be drawn on the screen cast stream.

types - Sets the types of content to record.

multiple- Sets whether to allow selecting multiple sources.

pub async fn start(
&self,
session: &SessionProxy<'_>,
identifier: WindowIdentifier
) -> Result<Vec<Stream>, Error>

Start the screen cast session.

This will typically result the portal presenting a dialog letting the user do the selection set up by select_sources.

An application can only attempt start a session once.

Arguments

session - A SessionProxy.

identifier - Identifier for the application window.

pub async fn available_cursor_modes(
&self
) -> Result<BitFlags<CursorMode>, Error>

Available cursor mode.

pub async fn available_source_types(
&self
) -> Result<BitFlags<SourceType>, Error>

Available source types.