Struct ashpd::documents::Documents

pub struct Documents<'a>(_);

The interface lets sandboxed applications make files from the outside world available to sandboxed applications in a controlled way.

Exported files will be made accessible to the application via a fuse filesystem that gets mounted at /run/user/$UID/doc/. The filesystem gets mounted both outside and inside the sandbox, but the view inside the sandbox is restricted to just those files that the application is allowed to access.

Individual files will appear at /run/user/$UID/doc/$DOC_ID/filename, where $DOC_ID is the ID of the file in the document store. It is returned by the Documents::add and Documents::add_named calls.

The permissions that the application has for a document store entry (see Documents::grant_permissions) are reflected in the POSIX mode bits in the fuse filesystem.

Wrapper of the DBus interface: org.freedesktop.portal.Documents.

Implementations

impl<'a> Documents<'a>

pub async fn new() -> Result<Documents<'a>, Error>

Create a new instance of Documents.

pub async fn add( &self, o_path_fd: &impl AsRawFd + Debug, reuse_existing: bool, persistent: bool ) -> Result<DocumentID, Error>

Adds a file to the document store. The file is passed in the form of an open file descriptor to prove that the caller has access to the file.

Arguments

o_path_fd - Open file descriptor for the file to add.

reuse_existing - Whether to reuse an existing document store entry for the file.

persistent - Whether to add the file only for this session or permanently.

Returns

The ID of the file in the document store.

Specifications

See also Add.

pub async fn add_full( &self, o_path_fds: &[&impl AsRawFd], flags: BitFlags<DocumentFlags>, app_id: Option<AppID>, permissions: &[Permission] ) -> Result<(Vec<DocumentID>, HashMap<String, OwnedValue>), Error>

Adds multiple files to the document store. The files are passed in the form of an open file descriptor to prove that the caller has access to the file.

Arguments

o_path_fds - Open file descriptors for the files to export.

flags - A DocumentFlags.

app_id - An application ID, or None.

permissions - The permissions to grant.

Returns

The IDs of the files in the document store along with other extra info.

Specifications

See also AddFull.

pub async fn add_named( &self, o_path_parent_fd: &impl AsRawFd + Debug, filename: impl AsRef<Path>, reuse_existing: bool, persistent: bool ) -> Result<DocumentID, Error>

Creates an entry in the document store for writing a new file.

Arguments

o_path_parent_fd - Open file descriptor for the parent directory.

filename - The basename for the file.

reuse_existing - Whether to reuse an existing document store entry for the file.

persistent - Whether to add the file only for this session or permanently.

Returns

The ID of the file in the document store.

Specifications

See also AddNamed.

pub async fn add_named_full( &self, o_path_fd: &impl AsRawFd + Debug, filename: impl AsRef<Path>, flags: BitFlags<DocumentFlags>, app_id: Option<AppID>, permissions: &[Permission] ) -> Result<(DocumentID, HashMap<String, OwnedValue>), Error>

Adds multiple files to the document store. The files are passed in the form of an open file descriptor to prove that the caller has access to the file.

Arguments

o_path_fd - Open file descriptor for the parent directory.

filename - The basename for the file.

flags - A DocumentFlags.

app_id - An application ID, or None.

permissions - The permissions to grant.

Returns

The ID of the file in the document store along with other extra info.

Specifications

See also AddNamedFull.

pub async fn delete(&self, doc_id: impl Into<DocumentID>) -> Result<(), Error>

Removes an entry from the document store. The file itself is not deleted.

Note This call is available inside the sandbox if the application has the Permission::Delete for the document.

Arguments

doc_id - The ID of the file in the document store.

Specifications

See also Delete.

pub async fn mount_point(&self) -> Result<FilePath, Error>

Returns the path at which the document store fuse filesystem is mounted. This will typically be /run/user/$UID/doc/.

Specifications

See also GetMountPoint.

pub async fn grant_permissions( &self, doc_id: impl Into<DocumentID>, app_id: AppID, permissions: &[Permission] ) -> Result<(), Error>

Grants access permissions for a file in the document store to an application.

Note This call is available inside the sandbox if the application has the Permission::GrantPermissions for the document.

Arguments

doc_id - The ID of the file in the document store.

app_id - The ID of the application to which permissions are granted.

permissions - The permissions to grant.

Specifications

See also GrantPermissions.

pub async fn info( &self, doc_id: impl Into<DocumentID> ) -> Result<(FilePath, Permissions), Error>

Gets the filesystem path and application permissions for a document store entry.

Note This call is not available inside the sandbox.

Arguments

doc_id - The ID of the file in the document store.

Returns

The path of the file in the host filesystem along with the Permissions.

Specifications

See also Info.

pub async fn list( &self, app_id: Option<AppID> ) -> Result<HashMap<DocumentID, FilePath>, Error>

Lists documents in the document store for an application (or for all applications).

Note This call is not available inside the sandbox.

Arguments

app-id - The application ID, or None to list all documents.

Returns

HashMap mapping document IDs to their filesystem path on the host system.

Specifications

See also List.

pub async fn lookup( &self, filename: impl AsRef<Path> ) -> Result<Option<DocumentID>, Error>

Looks up the document ID for a file.

Note This call is not available inside the sandbox.

Arguments

filename - A path in the host filesystem.

Returns

The ID of the file in the document store, or None if the file is not in the document store.

Specifications

See also Lookup.

pub async fn revoke_permissions( &self, doc_id: impl Into<DocumentID>, app_id: AppID, permissions: &[Permission] ) -> Result<(), Error>

Revokes access permissions for a file in the document store from an application.

Note This call is available inside the sandbox if the application has the Permission::GrantPermissions for the document.

Arguments

doc_id - The ID of the file in the document store.

app_id - The ID of the application from which the permissions are revoked.

permissions - The permissions to revoke.

Specifications

See also RevokePermissions.