Struct ashpd::documents::DocumentsProxy

pub struct DocumentsProxy<'c>(_);

The interface lets sandboxed applications make files from the outside world available to sandboxed applications in a controlled way.

Exported files will be made accessible to the application via a fuse filesystem that gets mounted at /run/user/$UID/doc/. The filesystem gets mounted both outside and inside the sandbox, but the view inside the sandbox is restricted to just those files that the application is allowed to access.

Individual files will appear at /run/user/$UID/doc/$DOC_ID/filename, where $DOC_ID is the ID of the file in the document store. It is returned by the Add() and AddNamed() calls.

The permissions that the application has for a document store entry (see GrantPermissions()) are reflected in the POSIX mode bits in the fuse filesystem.

Implementations

impl<'c> DocumentsProxy<'c>

pub fn new(conn: &Connection) -> Result<Self>

Creates a new proxy with the default service & path.

pub fn new_for(
conn: &Connection,
destination: &'c str,
path: &'c str
) -> Result<Self>

Creates a new proxy for the given destination and path.

pub fn new_for_owned(
conn: Connection,
destination: String,
path: String
) -> Result<Self>

Same as new_for but takes ownership of the passed arguments.

pub fn into_inner(self) -> Proxy<'c>

Consumes self, returning the underlying zbus::Proxy.

pub fn inner(&self) -> &Proxy<'_>

The reference to the underlying zbus::Proxy.

pub fn add(
&self,
o_path_fd: Fd,
reuse_existing: bool,
persistent: bool
) -> Result<String>

Adds a file to the document store. The file is passed in the form of an open file descriptor to prove that the caller has access to the file.

Returns the ID of the file in the document store.

Arguments

o_path_fd - open file descriptor for the file to add

reuse_existing - whether to reuse an existing document store entry for the file

persistent - whether to add the file only for this session or permanently

pub fn add_full(
&self,
o_path_fds: &[Fd],
flags: BitFlags<Flags>,
app_id: &str,
permissions: &[Permission]
) -> Result<(Vec<String>, HashMap<String, OwnedValue>)>

Adds multiple files to the document store. The files are passed in the form of an open file descriptor to prove that the caller has access to the file.

Returns the IDs of the files in the document store along with other extra info.

Arguments

o_path_fds - open file descriptors for the files to export

flags - a Flags enum.

app_id - an application ID, or empty string

permissions - the permissions to grant

pub fn add_named(
&self,
o_path_parent_fd: Fd,
filename: &NString,
reuse_existing: bool,
persistent: bool
) -> Result<String>

Creates an entry in the document store for writing a new file.

Returns the ID of the file in the document store.

Arguments

o_path_parent_fd - open file descriptor for the parent directory

filename - the basename for the file

reuse_existing - whether to reuse an existing document store entry for the file

persistent - whether to add the file only for this session or permanently

pub fn add_named_full(
&self,
o_path_fd: Fd,
filename: &NString,
flags: BitFlags<Flags>,
app_id: &str,
permissions: &[Permission]
) -> Result<(String, HashMap<String, OwnedValue>)>

Adds multiple files to the document store. The files are passed in the form of an open file descriptor to prove that the caller has access to the file.

Returns the ID of the file in the document store along with other extra info.

Arguments

o_path_fd - open file descriptor for the parent directory

filename - the basename for the file

flags - a Flags

app_id - an application ID, or empty string

permissions - the permissions to grant.

pub fn delete(&self, doc_id: &str) -> Result<()>

Removes an entry from the document store. The file itself is not deleted. This call is available inside the sandbox if the application has the 'delete' permission for the document.

Arguments

doc_id - The ID of the file in the document store

pub fn get_mount_point(&self) -> Result<NString>

Returns the path at which the document store fuse filesystem is mounted. This will typically be /run/user/$UID/doc/.

pub fn grant_permissions(
&self,
doc_id: &str,
app_id: &str,
permissions: &[Permission]
) -> Result<()>

Grants access permissions for a file in the document store to an application. This call is available inside the sandbox if the application has the 'grant-permissions' permission for the document.

Arguments

doc_id - the ID of the file in the document store.

app_id - the ID of the application to which permissions are granted.

permissions - the permissions to grant.

pub fn info(&self, doc_id: &str) -> Result<(NString, Permissions)>

Gets the filesystem path and application permissions for a document store entry.

Returns the path of the file in the host filesystem along with the [Permissions]

Arguments

doc_id - The ID of the file in the document store

pub fn list(&self, app_id: &str) -> Result<HashMap<String, NString>>

Lists documents in the document store for an application (or for all applications).

Returns a HashMap mapping document IDs to their filesystem path on the host system

Arguments

app-id - The application ID, or '' to list all documents

pub fn lookup(&self, filename: NString) -> Result<String>

Looks up the document ID for a file. This call is not available inside the sandbox.

Returns the ID of the file in the document store, or '' if the file is not in the document store

Arguments

filename - A path in the host filesystem

pub fn revoke_permissions(
&self,
doc_id: &str,
app_id: &str,
permissions: &[Permission]
) -> Result<()>

Revokes access permissions for a file in the document store from an application. This call is available inside the sandbox if the application has the 'grant-permissions' permission for the document.

Arguments

doc_id - The ID of the file in the document store

app_id - The ID of the application from which permissions are revoked

permissions - The permissions to revoke.

pub fn version(&self) -> Result<u32>

version property

pub fn introspect(&self) -> Result<String>