Struct ashpd::documents::FileTransferProxy

pub struct FileTransferProxy<'a>(_);

The interface operates as a middle-man between apps when transferring files via drag-and-drop or copy-paste, taking care of the necessary exporting of files in the document portal.

Toolkits are expected to use the application/vnd.portal.filetransfer mimetype when using this mechanism for file exchange via copy-paste or drag-and-drop.

The data that is transmitted with this mimetype should be the key returned by the StartTransfer method. Upon receiving this mimetype, the target should call RetrieveFiles with the key, to obtain the list of files. The portal will take care of exporting files in the document store as necessary to make them accessible to the target.

Wrapper of the DBus interface: org.freedesktop.portal.FileTransfer.

Implementations

impl<'a> FileTransferProxy<'a>

pub async fn new(
connection: &Connection
) -> Result<FileTransferProxy<'a>, Error>

Create a new instance of FileTransferProxy.

pub fn inner(&self) -> &Proxy<'_>

Get a reference to the underlying Proxy.

pub async fn add_files<F: AsRawFd>(
&self,
key: &str,
fds: &[&F]
) -> Result<(), Error>

Adds files to a session. This method can be called multiple times on a given session. Note only regular files (not directories) can be added.

Arguments

key - A key returned by start_transfer().

fds - A list of file descriptors of the files to register.

Specifications

See also AddFiles.

pub async fn retrieve_files(&self, key: &str) -> Result<Vec<String>, Error>

Retrieves files that were previously added to the session with add_files(). The files will be exported in the document portal as-needed for the caller, and they will be writable if the owner of the session allowed it.

Arguments

key - A key returned by start_transfer().

Returns

The list of file paths.

Specifications

See also RetrieveFiles.

pub async fn start_transfer(
&self,
writeable: bool,
auto_stop: bool
) -> Result<String, Error>

Starts a session for a file transfer. The caller should call add_files() at least once, to add files to this session.

Arguments

writeable - Sets whether the chosen application can write to the files or not.

auto_stop - Whether to stop the transfer automatically after the first retrieve_files() call.

Returns

Key that can be passed to retrieve_files() to obtain the files.

Specifications

See also StartTransfer.

pub async fn stop_transfer(&self, key: &str) -> Result<(), Error>

Ends the transfer. Further calls to add_files() or retrieve_files() for this key will return an error.

Arguments

key - A key returned by start_transfer().

Specifications

See also StopTransfer.

pub async fn transfer_closed(&self) -> Result<String, Error>

Emitted when the transfer is closed.

Returns

The key returned by start_transfer().

Specifications

See also TransferClosed.