Module ashpd::flatpak::update_monitor

Monitor if there’s an update it and install it.

Examples

How to monitor if there’s a new update and install it. Only available for Flatpak applications.

use ashpd::flatpak::update_monitor::{
    UpdateInfo, UpdateMonitorProxy, UpdateOptions, UpdateProgress,
};
use ashpd::flatpak::{CreateMonitorOptions, FlatpakProxy};
use ashpd::WindowIdentifier;
use zbus::{fdo::Result, Connection};

fn main() -> Result<()> {
    let connection = Connection::new_session()?;
    let proxy = FlatpakProxy::new(&connection);

    let monitor = proxy.create_update_monitor(CreateMonitorOptions::default())?;

    monitor.connect_progress(move |p: UpdateProgress| {
        if p.progress == Some(100) {
            monitor.close()?;
        }
        Ok(())
    })?;

    monitor.connect_update_available(move |_: UpdateInfo| {
        monitor.update(WindowIdentifier::default(), UpdateOptions::default())?;
        Ok(())
    })?;

    Ok(())
}

Structs

AsyncUpdateMonitorProxy
Asynchronous sibling of UpdateMonitorProxy. The interface exposes some interactions with Flatpak on the host to the sandbox. For example, it allows you to restart the applications or start a more sandboxed instance.

UpdateInfo
A response containing the update information when an update is available.

UpdateMonitorProxy
The interface exposes some interactions with Flatpak on the host to the sandbox. For example, it allows you to restart the applications or start a more sandboxed instance.

UpdateOptions
Specified options on an update request.

UpdateProgress
A response of the update progress signal.

Enums

UpdateStatus
The update status.